requirements.txt
requests
pyTelegramBotAPI
python-dotenv
imgkit